+++ tests/example_programs_test.rs
@@ -49,3 +49,4 @@ example_test!(data_pipeline, "02_data_pipeline.ol");
 example_test!(algorithms, "03_algorithms.ol");
 example_test!(stdlib_showcase, "04_stdlib_showcase.ol");
 example_test!(text_processing, "05_text_processing.ol");
+example_test!(database, "06_database.ol");